move interpolation function to helpers
