+++ src/festim/helpers.py
@@ -287,3 +287,25 @@ if version.parse(dolfinx_version) > version.parse("0.9.0"):
     get_interpolation_points = lambda element: element.interpolation_points
 else:
     get_interpolation_points = lambda element: element.interpolation_points()
+
+
+def nmm_interpolate(f_out: fem.function, f_in: fem.function):
+    """Non Matching Mesh Interpolate: interpolate one function (f_in) from one mesh into
+    another function (f_out) with a mismatching mesh
+
+    args:
+        f_out: function to interpolate into
+        f_in: function to interpolate from
+
+    notes:
+    https://fenicsproject.discourse.group/t/gjk-error-in-interpolation-between-non-matching-second-ordered-3d-meshes/16086/6
+    """
+
+    dim = f_out.function_space.mesh.topology.dim
+    index_map = f_out.function_space.mesh.topology.index_map(dim)
+    ncells = index_map.size_local + index_map.num_ghosts
+    cells = np.arange(ncells, dtype=np.int32)
+    interpolation_data = fem.create_interpolation_data(
+        f_out.function_space, f_in.function_space, cells, padding=1e-11
+    )
+    f_out.interpolate_nonmatching(f_in, cells, interpolation_data=interpolation_data)
